requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
google-generativeai==0.8.3
python-dotenv==1.0.1
langchain-core>=0.1.26,<0.4
langchain-google-genai>=2.0.7
langchain-community>=0.0.24
chromadb==0.6.1
PyMuPDF==1.23.8
pdf2image==1.17.0 